update all files
