+++ src/FMM/fitting/fit_params.yaml
@@ -3,7 +3,7 @@ cmb:
     fit: True
     bound_min: -1
     bound_max: 1
-    init_average: 0.0001
+    init_average: 0
     init_std: 0.1
   Alens:
     fit: 1
@@ -14,22 +14,22 @@ cmb:
 
 dust:
   Ad:
-    fit: 0
+    fit: True
     bound_min: 0
     bound_max: 10
     init_average: 1
-    init_std: 0.01
+    init_std: 0.1
   alphad:
-    fit: 0
-    bound_min: 0
+    fit: True
+    bound_min: -10
     bound_max: 10
-    init_average: 1
-    init_std: 0.01
+    init_average: -1
+    init_std: 0.1
   betad:
-    fit: 0
+    fit: 1.54
     bound_min: 0
     bound_max: 10
-    init_average: 1
+    init_average: 1.5
     init_std: 0.01
 
 synchrotron:
